feat(web): show integrations in settings pane
