+++ web/src/views/SettingsView.test.ts
@@ -67,9 +67,8 @@ function mountSettings() {
       stubs: {
         Icon: true,
         AppButton: { template: '<button type="button" v-bind="$attrs"><slot /></button>' },
-        IntegrationsModal: {
-          props: ['open'],
-          template: '<div data-testid="integrations-modal-stub" :data-open="open" />',
+        IntegrationsPanel: {
+          template: '<div data-testid="integrations-panel-stub" />',
         },
       },
     },
@@ -191,7 +190,9 @@ describe('SettingsView integrations entry card (plan g2.1 / g2.2 / g2.4)', () =>
   it('source keeps integrations card outside settings load branches', () => {
     expect(src).toMatch(/data-testid="settings-integrations-card"/)
     expect(src).toMatch(/data-testid="settings-integrations-open"/)
-    expect(src).toMatch(/IntegrationsModal/)
+    expect(src).toMatch(/IntegrationsPanel/)
+    expect(src).not.toMatch(/IntegrationsModal/)
+    expect(src).not.toMatch(/router\.replace/)
     expect(src).toMatch(/query\.integrations/)
   })
 
@@ -206,20 +207,22 @@ describe('SettingsView integrations entry card (plan g2.1 / g2.2 / g2.4)', () =>
     w.unmount()
   })
 
-  it('keeps integrations card on 403 and opens modal stub', async () => {
+  it('keeps integrations card on 403 and navigates to the inline panel', async () => {
     apiMocks.getSettings.mockRejectedValue(Object.assign(new Error('forbidden'), { status: 403 }))
     const w = mountSettings()
     await flushPromises()
     expect(w.find('[data-testid="settings-denied"]').exists()).toBe(true)
     expect(w.find('[data-testid="settings-integrations-card"]').exists()).toBe(true)
-    expect(w.find('[data-testid="integrations-modal-stub"]').attributes('data-open')).toBe('false')
+    expect(w.find('[data-testid="integrations-panel-stub"]').exists()).toBe(false)
     await w.find('[data-testid="settings-integrations-open"]').trigger('click')
     await flushPromises()
-    expect(w.find('[data-testid="integrations-modal-stub"]').attributes('data-open')).toBe('true')
+    expect(w.find('[data-testid="integrations-panel-stub"]').exists()).toBe(true)
+    expect((w.vm as any).$router.currentRoute.value.query.integrations).toBe('1')
+    expect(w.find('[data-testid="settings-integrations-card"]').exists()).toBe(false)
     w.unmount()
   })
 
-  it('auto-opens integrations modal from ?integrations=1 then clears query', async () => {
+  it('renders integrations inline from ?integrations=1 and keeps query', async () => {
     apiMocks.getSettings.mockResolvedValue(SETTINGS)
     const i18n = createI18n({
       legacy: false,
@@ -237,16 +240,17 @@ describe('SettingsView integrations entry card (plan g2.1 / g2.2 / g2.4)', () =>
         stubs: {
           Icon: true,
           AppButton: { template: '<button type="button" v-bind="$attrs"><slot /></button>' },
-          IntegrationsModal: {
-            props: ['open'],
-            template: '<div data-testid="integrations-modal-stub" :data-open="open" />',
+          IntegrationsPanel: {
+            template: '<div data-testid="integrations-panel-stub" />',
           },
         },
       },
     })
     await flushPromises()
-    expect(w.find('[data-testid="integrations-modal-stub"]').attributes('data-open')).toBe('true')
-    expect(router.currentRoute.value.query.integrations).toBeUndefined()
+    expect(w.find('[data-testid="integrations-panel-stub"]').exists()).toBe(true)
+    expect(w.find('[data-testid="settings-integrations-card"]').exists()).toBe(false)
+    expect(w.find('[data-testid="settings-form-skeleton"]').exists()).toBe(false)
+    expect(router.currentRoute.value.query.integrations).toBe('1')
     w.unmount()
   })
 })
